chore: rename transformer2d to transformer_2d
