+++ pipefuser/models/naive_patch_dit.py
@@ -1,6 +1,6 @@
 from pipefuser.models.base_model import BaseModel
 import torch
-from pipefuser.modules.dit.patch_parallel.transformer2d import DistriTransformer2DModel
+from pipefuser.modules.dit.patch_parallel.transformer_2d import DistriTransformer2DModel
 from diffusers.models.transformers.transformer_2d import Transformer2DModelOutput
 
 # from diffusers.models.transformers.transformer_2d import Transformer2DModel
