+++ pipefuser/modules/dit/patch_parallel/__init__.py
@@ -1,4 +1,4 @@
-from .transformer2d import DistriTransformer2DModel
+from .transformer_2d import DistriTransformer2DModel
 from .attn import DistriCrossAttentionPP, DistriSelfAttentionPP
 from .conv2d import DistriConv2dPP
 from .groupnorm import DistriGroupNorm
